Disable Datastore checks
The check is failing in our CI clusters, because their cluster ID has too
many dashes, such as "ci-op-cvx98rqr-d1161-28xzd"

Full volume name then is
/var/lib/kubelet/plugins/kubernetes.io/vsphere-volume/mounts/[WorkloadDatastore]
00000000-0000-0000-0000-000000000000/ci-op-cvx98rqr-d1161-28xzd-dynamic-pvc-00000000-0000-0000-0000-000000000000.vmdk,
which, escaped by systemd, is over the limit of 255 characters.

This is temporary measure, we need to implement more strict cutting of
cluster IDs in Kubernetes.