Don't degrade cluster on connection error
Clusters that do not have correct credentials to vCenter should not get
degraded when vsphere-problem-detector cannot connect to it.

Instead, keep the cluster Availabe=true and only report a new metric +
alert on it.

Add the error message to Availabe=true condition, so it can be found
without digging through logs.