Fix propagation of check errors to Available condition
Pass overall check error to Available condition of Storage CR. The code was
already prepared to do so, just runChecks() must return the actual error
(or list of errors).


When at it, fix Collect() to actually return list of all checks instead of
empty []checkResult. As consequence, the detector will emit events on all
check success or failure.