Lets remove datastore length checks for now
With CSI it will become less relevant and only solution is basically
renaming the datastore(which breaks existing PV).

As for systemd, it just means that systemd is not managing those
mount point as units. But I don't think it is necessary in this case
because kubelet is already doing it.